+++ packages/cuba-react-ui/src/index.less
@@ -4,4 +4,8 @@
 @import "ui/table/DataTableFilterControlLayout";
 @import "ui/table/DataTableIntervalEditor";
 @import "ui/table/DataTableListEditor";
+@import "ui/form/InputNumber";
+@import "ui/form/EntityEditor";
+@import "ui/form/NestedEntitiesTableField";
+@import "ui/form/NestedEntityField";
 @import "style/palette";
