+++ packages/cuba-react-ui/src/ui/table/DataTableCustomFilterFields.tsx
@@ -1,0 +1,39 @@
+import {IntlShape} from 'react-intl';
+import {GetFieldDecoratorOptions} from 'antd/es/form/Form';
+import React, {ReactNode} from 'react';
+import {Form} from 'antd';
+
+export function getDefaultFieldDecoratorOptions(intl: IntlShape): GetFieldDecoratorOptions {
+  return {
+    initialValue: null,
+    rules: [
+      {
+        required: true,
+        message: intl.formatMessage({id: 'cubaReact.dataTable.validation.requiredField'})
+      }
+    ]
+  };
+}
+
+export function decorateAndWrapInFormItem(
+  children: ReactNode,
+  parentId: string,
+  // tslint:disable-next-line:ban-types
+  getFieldDecorator: <T extends Object = {}>(id: keyof T, options?: GetFieldDecoratorOptions | undefined) => (node: ReactNode) => ReactNode,
+  intl: IntlShape,
+  hasFeedback: boolean = false,
+  options?: GetFieldDecoratorOptions,
+  additionalClassName?: string,
+): ReactNode {
+  if (!options) {
+    options = getDefaultFieldDecoratorOptions(intl);
+  }
+
+  return (
+    <Form.Item hasFeedback={hasFeedback} className={`filtercontrol ${additionalClassName || ''}`}>
+      {getFieldDecorator(`${parentId}_input`, options)(
+        children
+      )}
+    </Form.Item>
+  );
+}
